Remove embedded hiera() function. Use Automatic Parameter Lookup.
1) The hiera() function is deprecated and replaced with lookup().

2) It's more portable to rely on Automatic Parameter Lookup to get data
from Hiera than to have the hiera() function in your code. Also, it's
not standard to have the hiera() function in a componenet module, which
this is.

3) Replaced create_resources with .each() and iteration using Puppet 4
language features. It's more idiomatic and is easier to debug and see
what's going on.

After these changes, the module functions exactly as it did before.
Hiera data can be set using the same syntax as shown in the README.